adding Exit as an alias of Quit (#998)
This fixes #996. It seems that the rule for `Exit` was not loaded
together with `Quit`.

---------
