+++ mathics/builtin/mainloop.py
@@ -222,56 +222,3 @@ class Line(Builtin):
 
     name = "$Line"
     summary_text = "current line number"
-
-
-class Out(Builtin):
-    """
-    <url>:WMA: https://reference.wolfram.com/language/ref/$Out</url>
-    <dl>
-      <dt>'Out[$k$]'
-      <dt>'%$k$'
-      <dd>gives the result of the $k$th input line.
-
-      <dt>'%', '%%', etc.
-      <dd>gives the result of the previous input line, of the line before the previous input line, etc.
-    </dl>
-
-    >> 42
-     = 42
-    >> %
-     = 42
-    >> 43;
-    >> %
-     = 43
-    >> 44
-     = 44
-    >> %1
-     = 42
-    >> %%
-     = 44
-    >> Hold[Out[-1]]
-     = Hold[%]
-    >> Hold[%4]
-     = Hold[%4]
-    >> Out[0]
-     = Out[0]
-
-    #> 10
-     = 10
-    #> Out[-1] + 1
-     = 11
-    #> Out[] + 1
-     = 12
-    """
-
-    attributes = A_LISTABLE | A_PROTECTED
-
-    rules = {
-        "Out[k_Integer?Negative]": "Out[$Line + k]",
-        "Out[]": "Out[$Line - 1]",
-        "MakeBoxes[Out[k_Integer?((-10 <= # < 0)&)],"
-        "    f:StandardForm|TraditionalForm|InputForm|OutputForm]": r'StringJoin[ConstantArray["%%", -k]]',
-        "MakeBoxes[Out[k_Integer?Positive],"
-        "    f:StandardForm|TraditionalForm|InputForm|OutputForm]": r'"%%" <> ToString[k]',
-    }
-    summary_text = "result of the Kth input line"
